[feature] cli convert: add filename expension to tile names
The --tiles argument can take many filenames. That can be a lot for the
command line. In some cases, it's also not handy to pass the whole list
(ex, when debugging, or running on Windows).

So detect that there is only one argument, and in this case, attempt to
expend it using the standard glob rules.
In the case it was a normal file, just that file is returned. So the
current behaviour is not changed, just "expended".