[fix] acq stiching registrar: fix detection of too large shift
Surprisingly, the global shift registrar and the shift registrar don't
compute the shift between tiles in the same way.
But both functions had issues in the code deciding whether the shift is
too large to be correct.

Typically the overlapping areas are very elongated. So the max distance
in X and Y are very different. The shift vector should be compared
independently in X & Y.
The global shift method had an especially bad bug mixing Y & X, causing
it to estimate the overlap sometimes very small, which caused it to
almost always discard the estimated shift.