[ci] fastem mocked test cases added to cloud hosted CI
In fastem_test.py do not import from fastem_calibrations at the beginning of the file, because then all test cases will fail that depend on fastem_calibrations.
Mock fastem_calibrations and fastem_calibrations.util, to be able to call them on systems where fastem_calibrations is not available.
It should be possible to run the test cases that are not dependent on fastem_calibrations on systems that do not have this repository.