[fix] AngularSpectrumSettingsStream update binning on acquisition start too
In the EK (AR spectrum) stream, If the user changed the binning settings, and immediately started an
acquisition, without playing the stream, then the binning would not be set.
(This is a rare case)

=> Instead of having a special function for connecting the binning VAs,
override the general _*linkHwVAs() functions. The acquisition MD streams
are aware of these functions and explicitly call them.

This even simplifies the code.