[SDEV-1745] feat: add fastem calibration data and image to odemis bug reporter
The fastem calibration store log and results in odemis-status dir. 
Add it to the bug report, as it could help with a quicker first diagnosis of the problem. Exclude the images directory, because that would take up too much space. Still store one image, the cell translation plot, gives information about the quality of the multiprobe pattern.

This assumes that the lastest fastem-calibrations module is used, as it the old version didn't store all the data in odemis-status (but in that case, nothing bad will happen, just not all the data will be in the bug report).