[feature] new plugin for SPARC spectrum acquisition with arbitrary scan order
Allow to acquire a spectrum acquisition where the beam doesn't scan in
the standard order (ie, fast X, then slow Y). The plugin scans by
default along a checkboard pattern, but the idea is that the user can
change the patter to his/her liking by adding a different function.