[NSRPARMA-16] Add test case for scan-stage actuator wrapper
Addition of the scan-stage wrapper test case in stream_test as a
separate class. Consists of testing of the wrapper through a big
configuration file containing the new wrapper actuator type.
In addition contains a method to test for ccd-type detector only.
Also made the roi_to_phys method a global one to use for other
test cases.