[fix] Extend andorshrk to detect if flipper failed to move
Some hardware have sometimes issues with the flipper, and fail to move
it. In such case the move would "succeed" and the position left as the
previous position.

Instead, try to move again, which seems to work around some hardware
issues, and if that didn't help report an error with the move.