[feature] New SPARC Mirror actuator interlock plugin
Check for interlock affecting the SPARC parabolic mirror. If the interlock is triggered, the
"ENGAGE MIRROR" button is disabled. If the mirror is engaged when the interlock is triggered, the
mirror is automatically parked.

Obviously, it needs a interlock. So add a simulated interlock to a SPARC
simulator. Activate the interlock with:
odemis set-attr bsd-interlock interlockTriggered True