[refactor] move all MIRROR_* constants from sparc2_chamber_tab back to tabs
The MIRROR_* constants are used by the "sparc_engage_warn" plugin, and
soon by another one.
We could change the imports, but it's nicer to put them back to where
they were so that we don't break old plugin versions. Also, it seems
to make more sense to have all MIRROR_* constants in the same place.