[feature] driver andorshrk: add support for controlling the iris and input flipper
The Kymera 328 supports a diaphragm iris, to control how much light comes in.
Add ability to control it if it's present.

The hardware unit is arbitrary (0->100). We convert it to something closer to reality.

Also introduce support for the input flipper (flip-in) axis, as the
Kymera 328 as such motorized input, and it's important to be able to
select the right input! Mostly just copy-paste from the flip-out axis.