[fix] autostig period for idx 0 behaved incorrectly
For idx 0 even if the autostig period was set to 0 it would still run the extended pre-calibrations. Additionally it should not run for idx 0, only for subsequent indices.
Re-initialize AcquisitionConfig() instead of using conf.get_acqui_conf() for autostigmation and for the overlap, so the GUI does not have to be restarted when changing the value. This increases flexibility during testing.