[fix] driver simcam: on long exposure simulate properly over-exposure
The Simcam simulates the exposure by multiplying the data by the
relative change of exposure. That means that having an exposure time
twice longer returns values in the data twice bigger. However, on ints,
it could overflow, and return very small values, instead of clipping
to the maximum value the sensor accepts. This would cause "rainbow
effect" in the simulated images with long exposure time.

=> Clip to the data type maximum. (code copied from the img.Bin()
function).