[fix] fastem save full cell images, set values after calibration values are applied.
The values to save the full cell images were set before the calibration values were applied to the detector.
This means that if calibration 3 was run for the current scintillator the cell translations were first set to 0 to enable saving the full cell images.
Then it was set to the calibrated value. This caused an error to be raised on the server side, about invalid cell parameters.