[feature] New pwrmccdaq driver for the Diligent MCC DAQ 1208LS
Add support for the Diligent MCC DAQ 1208LS (USB DAQ board with 2 AO).
This provides two interfaces: one to just read the digital inputs and export the state as a BooleanVA.  (Will be used for detecting interlock state).
The second interface allows to control the power of light/laser (similar to pwrcomedi). It can adjust the analog output voltage based on the light intensity requested. It also can turn on/off the light completely via a digital output, when the power is set to 0.

Relies on python3-hid and python3-usb1.