[fix] at startup, always show the components in the same order
Until now, the component shown in the back-end startup window were in random order. Ideally, we would show them in the same order they are started... but it turns out it's pretty hard to know in advance. So at least, show them in alphabetical order, so that each run looks similar. It's less confusing for the users.