[fix] driver andorcam2: force frameDuration update after every setting change
We need to know how long it will take for the hw sync, before we
actually acquire a frame. So make this VA proper, and immediately update
it when the affecting settings are changed.
Only do it when not acquiring, as it's not possible when acquiring...
and anyway in this case, it's unlikely we care about frameDuration (until next
frame).