[fix] dataio hdf5: don't warn it's not possible to save MD_WL_LIST if it's empty
It's not possible to remove metadata from the component. So if there is
no MD_WL_LIST, the best we can do is to make it None or a empty list.
So don't try to save the MD_WL_LIST if it's empty (ex, for AR data), and
so don't warn that it's not possible to store it.