[fix] Change odemisd.modelgen default to no strict syntax
For now, most of the microscope files are still incorrect. Defaulting to
strict syntax made the util.testing functions fail to parse the
microscope files.
We could switch that function alone, but it makes more sense to just
change the default, as long as most of the microscope files have errors.