[fix] andorcam2: support very high frame rate
If the frame rate > 100Hz, the camera sends multiple images in bursts
(every 10 ms).
Until now, the driver would only read one image per burst. Now properly
detect such case, and attempt to read as many as possible. (it can be
easier than it sounds, as typically, this can happen with very small
images, with a single line).

Also make sure the frame duration is updated when the synchonization
changes, as the hw trigger affects timings.