[fix] driver semnidaq: AO event buffer not called and incorrect DO period
Fix issue when the AO & DO buffers would be sligtly too small for the
6361, which caused the event callback from not being called. This meant
the AO & DO would not be pushed to the hardware, and it would starve.

=> Increase the minimum buffer size by 1, which seems to be enough (4095
to 4096). It's not very clear the relationship, and couldn't find any
mention of it in the doc, and no one answered on the forum... so let's
just keep this simple solution for now.

Also fix issue when DO period would be incorrectly accepted, due to a
copy-paste error.