[feature] model.DataFlow automatically keep all data when synchronized
Introduce a way to have the .max_discard attribute of the DataFlow be
directly seen by the DataFlowProxy.

With this functionality in place, it's now really worthy to adjust
max_discard when the DataFlow is synchronized, on the backend side
(only).

So introduce a generic .synchronizedOn() method on DataFlow that all
inheriting DataFlows (implemented by the drivers) can use. A lot of
drivers actually already did this, so this mostly simplifies the code.
The main change is that now the proxy really will NOT discard data
when the DataFlow is synchronized (as most of the discard happend on the
client side).

Note that this now "upgrade" the code to only support 0MQ v3+. That is
available since Ubuntu 16.04, so should really not be an issue!