[feature] spectrometer wrapper force settings when synchronized
In order to know the duration a frame will take, the camera component
must have all the VAs set. However, normally the spectrometer driver
only sets the VA just before the DataFlow is started. That's too late.
So, as a workaround, at least force the settings "flush" to also happen
when calling "synchronizedOn".

This at least allows to support our main use case of frameDuration,
which is to know the duration of the CCD while it's synchronized with
the e-beam scanner.