[fix] dataio hdf5: don't warn it's not possible to save MD_WL_LIST if it's empty
MD_WL_LIST is set on the spectrometer components, and updated whenever the specrograph moves, by the mdupdater. It uses .updateMetadata(). It's not possible to remove metadata from the component using this function, just update it.

So when the spectrograph is not in a "1st order diffraction" mode (for example, in mirror mode), there is no wavelength information. Ideally there would be no MD_WL_LIST, but instead, the best we can do is to make it None or a empty list.

So if the MD_WL_LIST is empty, it just means the same as it's missing. In such case, let's not try to save the MD_WL_LIST (ex, for AR data), and especially don't scare users by warning that it's not possible to store it.