[fix] driver tmcm: round the axes limits to a multiple of ustepsizes
If the axis range is specified, moving precisely to the limit might end
up with a move which is slighly outside of the range (because the move
is rounded to a number of stepsizes).

Note that there are other reasons for the position to be outside of the
range limit, and so clients should not rely on the position to always be
within stage limit. However, in this case, it's convenient to report
precisely the minimum and maximum position the axis can reach.

=> Round the axis limits the same way as the position is rounded.