[config] better example file for the time correlator
The OD filter actually starts from 0.04 (instead of 0).
Also, now that the tmcm driver takes care of the rounding, no need to do
it in the microscope file.

Also remove the children from the microscope, which are not needed
anymore.
Also add range to the Optical Actuators, as we do for newer SPARCs.