[fix] "Stop all axes" action not working
The microscope doesn't have "children" anymore, since commit 37456f109e33
(Do not require that all the components be listed as children of the
microscope).
However, this commit also stopped adding all the components to
.children. Now the components are only accessible via the "alive" VA.

Everywhere, .alive is used to list the components (usually indirectly,
via model.getComponents())... except in the case of the GUI "Stop all
Axes" function. So the function couldn't find any component, and had no effect.
=> Also use .alive