[feat] fastem: reference the stage before overview image as well as after reloading the sample holder
The stage was not referenced after reloading a sample, this caused the user to see a misalignment between the live view and the overview image.
To make sure this does not happen again we will automatically reference the stage after reloading (and the chamber is thus in vacuum) the sample holder and before acquiring an overview image.