[performance] unbind shape events and handle those by ShapesOverlay, optimize _get_shape method by making use of cKDTree.
The cKDTree has data about the shape's center. When a on_left_down event occurs find 4 shapes whose centers are closest to the event position. Out of the 4 shapes check if the event position happened inside the shape and if that shape was selected by the user.
By doing so avoid iteration over all the shapes.
The shapes were drawn 2 times, this has been fixed by only allowing ShapesOverlay draw the shapes. A newly created shape is not added to the world overlay.