[fix] fastem overlap 0 error after failed pre-calibrations
To determine the position of where to run the pre-calibrations, the overlap between fields is set to 0. This is allowed for determining the position, but on the hardware an overlap of 0 is not allowed.
When the pre-calibrations failed the value was not set back to the initial overlap.
This was only a problem at the start of a new acquisition, because then the overlap is actually set on the hardware.
This fix adds a finally to the try-except around the pre-calibrations, to set back the overlap. To do this the code to prepare the pre-calibrations was moved to inside the precalibrate function.

The following error was raised in case of a fail:
2024-05-16 08:43:56,881	ERROR	log:46:	Remote exception Traceback (most recent call last):
  File "/usr/lib/python3/dist-packages/Pyro4/core.py", line 861, in handleRequest
    data=method(*vargs, **kwargs)   # this is the actual method call to the Pyro object
  File "/home/patel/development/odemis/src/odemis/driver/technolution.py", line 1746, in updateMetadata
    raise ValueError(f"Overlap of 0 is not allowed, suggested overlap: {suggested_overlap * 100}%")
ValueError: Overlap of 0 is not allowed, suggested overlap: 0.5000000000000004%