[fix] properly disconnect from stream panels when a window is closed
When a window containing a list of streams (panels) is closed, the
stream bar.clear() is called to disconnect all the streams.
Commit b075a230a (gui: remove pubsub dependency) broke this behaviour
in .clear() as this function first unbind the DESTROY event.

This explains the issues with the change of tint that commit ca48bbf9e1
(fix C/C++ wrapper runtime error when switching colormap) tried to fix.

Now also call the destroy callback in .clear().

=> This means we don't need to specifically disconnect .tint in the
stream panel. So this code can be removed. This code also had introduced
an issue when the stream has no .tint attribute (because it didn't use
hasattr), so it's even better to not need it.