[refactor] Use acquisition config for FAST-EM spot estimation threshold, not hard coded value.
The value used for the treshold of the algorithm used in estimate_grid_orientation_from_img was hardcoded everywhere.
During development and for very thick samples an expert user or developer might want to change this, this was for instance necessary in the PSI testing.
With this refactor makes it possible to update the value in the acquisition.ini file, instead of hardcoding it.