[fix] cli live view failed to resize image if window size is limited by screen size
Normally the window size is set to be precisely the image size. So the
magnification is 1, and there is actually no resizing of the image that
happens.

However, if the screen is smaller than the image size, the window is
limited by the screen size... and the magnification is a float < 1.
This caused the resize() input argument to be a float, while it expects
a number of pixels as int. In this case, no image is displayed.

=> Force conversion to an int.