[refactor] Remove GUI dependency of fastem acquisition and clean up imports
To have a default value for the ROC color the fastem acquisition module dependent on `odemis.gui.FG_COLOUR_WARNING`.
Ideally, nothing from odemis.acq depends on odemis.gui. Since FG_COLOUR_WARNING was only used as a default for an input argument, and in practice it is overwritten the code was refactored to use the string value of FG_COLOUR_WARNING as default.

Additionally the import order of odemis.acq.fastem has been refactored to match the guidelines.