[fix] make nfpm driver work with NewFocus v3
The NewFocus v3 doesn't support moving multiple axes simultaneously.
It's not clear whether the v2 (that we used before) did support it, or
also didn't but we just never noticed (which would be odd because several test cases do
this).

=> just serialize the move of each axis if move of several axes are
requested simultaneously. In practice, we never need to move several
axes simultaneously, so it's not a limitation.