[fix] MultiplexActuator shortcuts Future to the sub-component also with multiple axes
The wrapper has a "shortcut" to not create a dedicated Future for each
move if there is only one axis, because in that case the sub-component
can directly take care of synchronizing the moves. However, the rule can
be slightly extended to multiple axes, as long as they are all handled
by the same sub-component.