[feature] driver semnidaq: extend active state to be available on detectors too
For the counting PMT, we need to have a way to turn it on/off explicitly
via a TTL port whenever we are acquiring.
It's essentially the same logic as for the scanner. So generalise the
scanning state TTLs manager into a dedicated class, and introduce two
arguments to be able to use it on the Detectors.