[fix] andorshrk: don't fail when requested monochromator wavelength info with 0th order
When a monochromator is attached to the spectrograph, the grating and slit opening
defines which wavelength are observed. However, when the wavelength is
at 0 (0th order), then there is no wavelength information (because all
the light is transmitted. Instead of completly failing, just report a
very large range, to indicate whole the light is passed.