[feature] extend PCU driver to handle several PCUs connected
So far, the Power Control Unit (PCU) driver assumed only one PCU at most was connected
to the computer. To support systems with multiple PCUs,
we need to be able to distinguish them. Still, each instance of the
driver will control just one PCU.

As PCUs don't have serial number, instead we use the IDs of the EEPROMs
which are connected to the PCU. To handle cases where a single
microscope file is used in multiple configurations (so different
EEPROMs), we also introduce a way to *exclude* some IDs.

=> if IDs are not found, look for more devices. If "exclude_ids" are
provided, skip devices with any of these IDs.