[fix] andorcam2: always update settings in the same thread
With commit a151eb0d6 (force frameDuration update after every setting
change), the camera settings can be changed even when the camera is not
about to acquire. In this case, the settings are changed directly from
the "client" thread. If just after an acquisition is started, the
settings might be changed again from the acquisition thread. In some
(very rare) cases, this can lead to crash in the SDK library.

So instead of immediately updating the settings, we now request the
acquisition thread to do it. This way, settings are set always from the
same thread, serially.

Also be extra careful not to open the shutter when just updating the
settings.