[feature] new DelayConnector Component wrapper
Just for providing a streak-delay component based on the
keysigth.Trueform component, in order to match the conventional
structure of components, as expected by Odemis GUI.

It also takes care of updating the delay whenever the streak-unit time
range changes.

Also move the responsability of updating the MD_TRIGGER_RATE to the
mdupdater, instead of the streak-ccd having to do it explicitly.
(Though it still does it if the delaybox is controlled via remoteex,
to make sure it's up to date)