[fix] improve default settings for the SPARC autofocus
Use more vertical binning, to get more signal. This allows to have a
shorter exposure time (so shorter procedure), and even still have 4x
more signal, handy for systems wher the signal is very low (like on the
SPARCs with external spectrographs).

The vertical resolution doesn't really matter for spectrograph focus, so
it's fine to reduce it.