[feature] introduce a andorshrk.ShamrockBus to access multiple spectrographs
The SDK library always claims all the spectrographs connected to the
system. If need to connect to multiple spectographs, the standard way
is to connect each of them from a separate process. However, as soon as
the first process is started, the SDK automatically claims all the
spectrographs. So the next processes are unable to access the other
spectrographs. For now the work around has been that the spectrograph
power is turned on just before connecting to it, so the first process
could not claim it. However, this is quite unreliable/slow, and it
doesn't work if some of the spectrograph power cannot be remotely
controlled.

=> Special class to instantiate all the spectrograph from a single
process.