[ci] Add cloud hosted CI on Ubuntu 24.04
The unit tests do not pass yet on Ubuntu 24.04. This ci-workflow can be used to
figure out what we need to test and verify fixes. It will not send a notification for a failed run,
but if you click on the specific github action you will see a failure for the specific build.
I created a separate workflow, because it is not possible to integrate in the regular workflow
such that you get a notification for failures on Ubuntu 20/22 but not on Ubuntu 24.