[fix] semnidaq: increase scheduler priority
Try to increase the scheduler priority of the whole process, in the hope
that this reduces the chances the hardware buffer is full before the SDK
reads it again.

There has been some acquisition failures that *might* be avoided by this
change. However, it's very hard to reproduce, and so also very hard to
confirm that this really helps. At least no ill effect has been noticed.