[fix] semnidaq: avoid unefficient cases when AI data needs a tiny bit more than 1 buffer
The AI buffer is normally adjusted in size to match the AI data.
However, if the duration of a single pixel is just a tiny bit more than
the default buffer duration (0.1 s), the buffer would end up fitting
just a tiny bit less that a complete pixel. As the default spot mode in
the GUI uses a dwell time of 0.1s, this would actually end-up quite
often that a whole frame consists of 1 buffer + a few bytes. These few
bytes end up costing a lot of time (relatively).

=> Detect such case and then increase the buffer size to read precisely
a whole frame.