[refactor] future.moves.urllib -> urllib
The "future" module is provided by the "past" package, which we had
already removed the dependency on (but forgot about these imports).
This was needed to support Python 2. So this can now be safely dropped.

The "bonus" is that the past package on Ubuntu 24.04 is broken, as it
relies on "imp" which is removed on Python 3.12. So this should help
making Odemis compatible with the latest Ubuntu.