[feature] andorshrk: support for changing turret on KY328i
The Kymera 328i requires new functions to move the turret
physically to a dedicated position for exchanging it.

Add these functions, which are only available in the latest version of
the SDK.

Also add version information functions, which have been introduced in
the latest version of the SDK.

Finally, introduce a mechanism to detect dynamically when the SDK
does support functions.