[refactor] No need to override inspect.getmembers() from Python 3.11
inspect.getmembers() in Python 3.10 and prior has an issue when __getattr__ is modified.
However, this is fixed in v3.11. So we can just use the standard code
(and avoid the risk of using an old version of getmembers() which is not
compatible with the newer Python anymore).

Also adjust the test case to not complain if Python works fine.