[fix] andorcam2 would sometimes detect images are missing if coming at ~100Hz
If the images are coming at ~100Hz, sometimes 2 images arrive at the
same time, and sometimes not. In such case, GetNumberNewImages() would
report a second image is available, and after reading it, WaitForAcquisition() would
report that same image again, but as it's already read, trying to read
it a second time would fail. It would be logged as:
"trying again to acquire image after error Call to GetNumberNewImages"

=> Call WaitForAcquisition() in non-blocking mode, so that if it
detects the image, it resets the flag, and doesn't cause the next call
to report an image.