[fix] StreamSettingsConfig should accept applying settings even if not all VAs are defined
So far, when calling apply_settings(), as soon as one of VA of the
stream is not defined in the settings entry, none of the other VAs
would be updated.

Change it so that all the settings defined are applied. It's OK that
some stream VAs are unknown.

Also adjust the test cases to always delete the test settings file on the computer,
as otherwise an old version of the file could prevent the test cases
from passing.