[feature] Provide a way in the SPARC2 alignment tab to override the default settings for the focus streams
We have a "clever" function that can find good settings for the focus
stream *most* of the time. However, due to the large variations in the
SPARCs (and especially the cameras), there are a few cases which would
work better with different settings.

To handle such SPARCs, allow the service engineer to define a file
~/.config/odemis/focus_settings.json with the name of the camera, and
the "good" settings to use.
An example of such file content:
[
{
    "name": "VisNIR Camera",
    "detExposureTime": 0.01
}
]