[fix] SPARC2 FSLT focus modes using the correct components
The manual focus of the streak cam would always turn on the brightlight,
while it might need the brighlight-ext.
=> change SPARC2 focus functions to automatically pick the correct
brightlight. That means the manual focus doesn't need to pass the
brightlight explicitly anymore.

The manual focus of the streak cam would also move always many axes
related to the specotgraph, even if the streak cam is on a separate
spectrograph => introduce the streak-focus-ext mode.

Also, pressing the "manual focus" button could block the GUI,
while the spectrograph is moving to wavelength 0.
=> rely on the optical path manager to switch to wavelength 0.

When manual focus is active, and switching mode, the brightlight would
stay on. => explicitly turn off the brightlight.

Also fix the OPM test cases.