[fix] piezomotor simulator finish moving properly
The move function was returned before the the move was finished, this caused the position to not be updated correctly.
And that caused a fail of TestFastEMCalibration.test_optical_autofocus. The binary autofocus iteratively moves to the right and left of a center position.
It could happen that after a few iterations the left position did not update, making left and right equal. The binary search algorithm then incorrectly assumed it had reached the minimum stepsize and the focus was found.