[fix] semnidaq scan garbled due to incorrect AO write/AI read
A sort of "last minute optimisation" had sneeked in, to make a tiny bit
faster and simpler the writing of AO data: instead of using a dedicated
writer, directly use ao_task.out_stream.write().

It turns out that the dedicate writer does extra clever stuff, related
to reorganizing the data according to the hardware requirement. Without
using it, the scan positions would be garbled, with X&Y mixed up,
causing completely incorrect images.

Similarly, for the AI, using in_stream.read() works... as long as there
is a single channel. Otherwise the use of Reader is necessary.