[perf] optimize FASTEM RoA overlay display and interaction
Don't bind events to every single EditableShape. If there are more than 10, it starts slowing the GUI... and some users need 200+ shapes! Instead, and handle those by ShapesOverlay, and optimize by making use of cKDTree.
The drawback is that the cKDTree uses just the center of the shape, so in some corner cases, the user might not be able to select a shape, but then just clicking near its center first will work.

Also:
* correct variable name, v_center instead of center as it is represented in view coordinates
* is_point_in_shape() method now checks for hover as it covers edges, rotation point and shape's inside. So it's now renamed check_point_proximity()
* Avoid drawing each shape twice.