[feature] SPARC acq: new stream to acquire with an "independent" EBIC
Introduce a new type of "setting" stream, which shows the live view of
the indenpendent EBIC while also starting the e-beam scanning using a second
dataflow, and copy the metadata.

Extend the "standard" synchronized e-beam acquisition to support
independent EBIC. Mostly a matter of copying the .resolution and
.dwellTime beforehand, and copying the metadata of the e-beam data onto
the EBIC data afterwards.