[feat] use SAMPLE_CENTERS md to store all information about the supported sample carriers
A sample carrier has a number of scintillators. A scintillator is associated with a number.
Further a scintillator is described by its shape and center. Given the shape type, the
shape's dimensions are described by its size (rectangle) or radius (circle).
Currently two shape types are supported namely rectangle and circle.
center (X, Y) [m] of each scintillator is measured from the bottom-left position of the sample carrier