[fix] Update path for saving beam shift correction images
The images for the beam shift correction where saved as [image-dir]/beam-shift-correction/[timestamp]/[idx]_after.tiff
But because all the ROAs are created at once, the timestamp for all roas in a project was the same and thus the images were overwritten for subsequent ROAs.
This commit implements a fix that stores the images as [image-dir]/beam-shift-correction/[timestamp]/[roa name]/[idx]_after.tiff.