[fix] semnidaq driver fix corner-case AI and CI synchronized acquisitions
Acquisitions longer than ~1h10min (>8Gb samples) failed just because
the SDK cannot count such a large number of sample. We actually don't need the
SDK to count, as we already did it internally.
So, if the acquisition is long, let the AI acquire "continuously", and only limit the
outputs (AO + DO). For them, the number of samples shouldn't be an
issue, as the frequency is a lot lower.

Also, synchronized acquisitions with CI failed because the very last
sample wasn't generated. That's because, as CI tasks cannot use the base
clock, it uses the AO clock. To end the acquisition of the last sample,
the tasks needs one extra tick... which never came in as the AO task
automatically stoppped.
=> Extend the AO task by one sample in such case (by duplicating the
last output, so nothing different on the actual output).