[feat] driver keysight: allow to use the front panel when "powered off"
When the device receives any remote command, the front panel is locked.
In order to allow the user to try "fancy" settings, we now unlock it
while the "power" setting is set to "off". When the power is set to "on"
again, the settings are reset as defined by the VAs.

So that means it's not necessary to regularly read the settings of the
device. As either it's locked, or it's not but we don't care (because
the settings will be reset).