[fix] driver semnidaq: minimum AI period should depend on the number of AI channels
The minimum period was fixed, based on the ai_max_multi_chan_rate
property. However, that rate is the total rate, by summing all the channels.
So the minimum period must be multiplied by the number of channels.
Without doing that, the acquisition would fail when more than 2 AI channels
are used.

That said, trials seem to show that the value can be higher than the
"minimum period * channel". => Let's assume this value is correct for 2
channels, and linearly scale based on the number of channels.