[fix] driver semnidaq: don't consider dwell time adjustment during acquisition to be a "settings update"
Just before an acquisition with multiple detectors, the dwell time can be adjusted to fit the
number of detectors. The acquisition will use that updated value.
However, until now, if this happened, a message to the acquirer was sent
to indicate settings have changed... so the acquirer would stop after
the first frame, and restart, with the same settings.

=> avoid to stop/start, by not sending such message in this case.