[fix] xt_client: only consider a move is really completed when position is near the target position
On *some* TFS systems *sometimes* the move ends, but the position is not yet updated. The position eventually updates, within a few seconds. However, this confuses quite a lot of code, as first it seems the stage hasn't moved (so the images acquired are at the wrong place), and then finally it "moves", although no move was requested.
=> Wait until the position is updated to report that the move is complete.

In practice, this mainly fixes issues with tile acquisitions.

Co-authored-by: Éric Piel <piel@delmic.com>