[fix] driver smaract: don't wait twice in a row for end of the move after referencing
The MCS2 driver has a special option to move to a fixed position after
referencing. An indentation error caused the wait for end of that move to be
called even if the move wasn't done.

It didn't have much effect, just slowing down the referencing procedure
by a few milliseconds, and making the log quite confusing because it
seemed a second "ghost move" was done after referencing.