[fix] Arbitrary scan order not compatible with hardware synchronized acquisition
Now that the SPARC acquisition supports hardware synchronized
acquisition, we have to be explicitly careful in the arbitrary scan
order plugin, as it's not (generally) possible to combine both options.

Also report error if the user attempts to use the scan stage. In theory,
it could be compatible, but extra code would need to be written.