[fix] The state of some components was not saved in the HW settings metadata
The SettingsObserver only store the state of the components it is given.
This is defined by the MainGUIData.
On the SPARC, components with a role containing a number, like ccd0, or
spectrometer0 weren't added to the list of components to observe. These
components are very useful to obersve too.
=> Observe *every* component with a role, instead of only the ones which
are explicitly listed in the MainGUIData.