[fix] Adjust level of the notification message to show hardware errors on Ubuntu 22.04
On Ubuntu 22.04, Gnome doesn't show on the screen the notifications
of level "info" (or below). They are only shown if the user looks at the
notification tray.

Errors with the microscope hardware are typically very important (on a computer
whose main purpose is to control the hardware). So raise the errors to
"ERROR" level, and when an error is recovered, to warning.
This way, all the user should notice these errors better.