[feat] Protect streak-cam when unblanking e-beam
When the e-beam is unblanked (even when going from pulsed-blanked to
unblanked), the sample may generate a lot more light. This can be
dangerous for the streak-cam.

=> If this happens, protect the streak-cam.

Also extend the sparc2 tabs to listen to the "hardware protection"
request by the user, and pause the streams in such case, for extra
safety (and keeping the stream VA updates simple).