[fix] kymera-exchange-turret fall back to direct connection if component not available on current back-end
The script can connect to the spectrograph in two ways: either via the
running back-end (based on the role), or directly (based on a serial
number).

If a back-end is available, the script would only try to connect via the
back-end. However, it can be that the current back-end doesn't control
the specific spectrograph. In this case, it shouldn't fail, but just try
via the direct connection.

Note that this doesn't solve *every* case, as if the backend connects to
another spectrograph, it might have also claimed the spectrograph we are
interested in, because that's how the Andor SDK works. In such case, the
only option is to stop the back-end. It's not a big deal, as after
switching the turret, the back-end must be restarted anyway.