[fix] hamamatsurx: .resolution incorrectly read
It assumed that HWidth and VWidth would report the resolution when
binning = 1x1. However, the value reported includes the binning.
So if started with a binning != 1x1 (which is common as typically
binning 2x2 is used), then every resolution reported was half the size.

=> Force binning to 1x1 before reading the resolution.

Also read the "accepted" resolution after changing the binning, to be
certain to report the actual resolution.