[feat] hamamatsurx: detect if hardware component missing or incorrect
If a component is missing (ie, not connected) or the wrong one is
connected (eg, wrong sweep unit), the HPDTA software doesn't really
complain about it. It just logs a warning and continue starting with the
device in simulation. That's unlikely what the user wants.

=> Detect such a log and show a error message to the user, so that they
can fix the connection and try again quickly.