[fix] driver andorcam2 doesn't update frameDuration until first image
The frameDuration is updated by the "acquisition" thread. However,
until now, the acquisition thread was only started on the first time an
acquisition was started. So, just after starting the back-end the
frameDuration would be incorrect.
That's actually not common on real systems, as the camera is almost
always first used before doing something relying on frameDuration, but
on test cases, this is very common.

Also adjust the simulator with the hardware trigger, to better simulate
external triggers typically don't arrive immediately when starting an
acquisition.