[fix] SPARC2HwSyncTestCase.test_acq_spec failure
Adjust code to take into account that the simulator is not so good,
so the CCD images start coming before the e-beam is started (which
is not the case on the real hardware), and can cause some timeout
in the acquisition code.
So reduce the repetition and increase the exposure time, to reduce the
delay between getting the CCD and the SEM data.