[fix] SEMCCDMDStream HW sync acquisition fails to update live image
Due to copy-paste from the standard acquisition, we used to reset the
raw data of the live CCD stream just after receiving it... so the
thread in charge of the update would frequently fail to get any data.

We don't need to reset the CCD image, as it's never integrated.