[fix] hamamatsurx with synchroscan unit could report time scale values 1000x too big
Apparently, the unit prefix selection when using the synchroscan sweep
unit is different from the single sweep unit. It seems to always use ps,
even for larger time ranges like 3.6 ns, for which the single sweep unit
would use ns. This caused the time scale to be incorrect by a factor of
1000x.

=> Detect that it's a synchroscan unit and always report ps.

Also adjust the time scale factor to be done every time the time scale
is read. This way, even if the user changes the time range from the
HPDTA software, the unit will be correctly detected.