[fix] keysight: don't get too confused if device complains when turning it on
Depending on the previous settings (which are random, as the user might
have changed them), when setting the device to generate a square waveform,
it might complain that some parameters are incompatible (and reset
them).
The driver used to then give up, and be confused because it states the
device is not on, but internally think it's on.

=> If this happens, try again, to confirm it wasn't a big deal.
If still an error happens, keep the internal state (_is_powered)
synchronized with the .power VA.

It used to look like that:

2024-10-09 18:40:47,717	DEBUG	keysight:599:	Sending command 'sour1:func SQU\n'
2024-10-09 18:40:47,717	DEBUG	keysight:577:	Sending query 'SYST:ERR?\n'
2024-10-09 18:40:47,741	DEBUG	keysight:591:	Received answer -221,"Settings conflict; frequency forced duty cycle change"\n