[feat] Add plugin to load FAST-EM factory settings from a dropdown menu
During debugging or service actions it can be necessary to be able to load the factory settings of the descanner. This will make sure the pattern is visible on the diagnostic camera.
This plugin adds an option in the help dropdown that will show a pop-up before loading the factory settings.