[fix] Run image translation before and after running image rotation
For reasons unknown it can happen that the image translation pre-alignment does not fully center the pattern on the mppc.
The result is that when we try to run the image rotation alignment, not all squares are visible in the image and the fit fails (see image below).
By first running the image translation alignment we ensure that all squares are visible in the image.
But because the image rotation does not rotate the pattern exactly around the center of the pattern we will have to run the image translation again afterwards.
So the order will be: image translation → image rotation → image translation.