[fix] Avoid e-beam.external "flickering" during SPARC SEM/CCD stream acquisition
Commit f24d37d55 (Avoid e-beam.external "flickering" during SPARC SEM MD
stream acquisition) introduced the idea of forcing the external mode
during acquisition. It turns out that the SEM CCD streams need the same.

So a similar code in this type of stream as well. This ensures that the
SEM stays in exernal mode all the time, and also reduces the acquisition
time, as sometimes switching external mode can take quite a long time
(ex: 0.5s).