[fix] SPARC hw sync acquisition didn't record properly polarimetry metadata
Between each polarization mode, the CCD should stop, like the SEM, and
restart from 0 for the next polarization. However, only the SEM
detectors were stopped, not the CCD one, which was stopped only at the
very end. As it's hardware synchronized, it almost work... but the CCD
driver only reads the metadata when starting to wait for the trigger.
So the metadata corresponded to the metadata for the previous
polarization, which caused polarimetry acquisitions to not display
properly.

=> Explicitly unsubscribe from the CCD data between each polarizaitions.