[fix] GUI shows checkbox unaligned
For the FASTEM GUI vEM (commit 42efe0cb6a), the checkbox position
in the size was changed. It seems this was to accomodate better the
calibration buttons.
However, this breaks the checkbox position on the SPARC (and METEOR)
streams.
=> Put back the checkbox positioning as before.

If the FASTEM GUI needs larger calibration buttons, then these buttons
should have the positionning adjusted.