[fix] pwrmccdaq: interlock could sometimes be detected incorrectly
It seems the low level USB connection is not thread-safe.
The interlock detection runs in a dedicated thread at 10Hz.
It appears that if during the read of the interlock DI status, the
light power was changed (on<>off), then, in some cases, they would
interfer with each other, and potentially cause the interlock reading to
report interlock (while it's not present).

=> Take a lock before accessing the device to ensure correct DI reading.