[fix] la_spec: also move lens 2 when using hardware triggered detectors
The lens 2 move happened in _adjustHardwareSettings(). However, when the
detectors can be hardware triggered, _adjustHardwareSettingsHwSync()
is called. So the lens 2 wasn't moved in such case. Similarly, if using
the scan stage, another function is called.

Instead of trying to override all these functions, do the same as the
LASpectrumSettingsStream: override the _prepare_opm(). It's actually
more logical, and should work in all cases.

Also avoid providing the stream on spectrographs which are not after the
lens 2. In such case, the acquisition would work... but it wouldn't be
a large FoV.